requirements.txt
flask==2.3.3
pillow==10.0.0
opencv-python==4.8.0.76
numpy==1.25.2
werkzeug==2.3.7
python-dotenv==1.0.0